# [Core] Avoid PENDING job to be set to FAILED and speed up job scheduling