--- a/sky/skylet/events.py
+++ b/sky/skylet/events.py
@@ -65,7 +65,6 @@ class JobSchedulerEvent(SkyletEvent):
 
     def _run(self):
         job_lib.scheduler.schedule_step(force_update_jobs=True)
-        logger.info('Triggerred job status update and scheduling.')
 
 
 class ManagedJobUpdateEvent(SkyletEvent):