--- a/tests/test_yamls/pipeline_aws.yaml
+++ b/tests/test_yamls/pipeline_aws.yaml
@@ -5,7 +5,7 @@ name: a
 
 resources:
   cloud: aws
-  region: us-west-2
+  region: us-east-2
   cpus: 2+
 
 run: |